// apps/allystudio/src/types/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      Page: {
        Row: {
          created_at: string
          deleted_at: string | null
          id: string
          normalized_url: string
          path: string
          updated_at: string
          url: string
          website_id: string
        }
        Insert: {
          created_at?: string
          deleted_at?: string | null
          id?: string
          normalized_url?: string
          path: string
          updated_at?: string
          url: string
          website_id: string
        }
        Update: {
          created_at?: string
          deleted_at?: string | null
          id?: string
          normalized_url?: string
          path?: string
          updated_at?: string
          url?: string
          website_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "Page_website_id_fkey"
            columns: ["website_id"]
            isOneToOne: false
            referencedRelation: "Website"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}